LPS-32615 Replace getScopeGroupId() with getSiteGroupId()
What is really needed in this class is always the groupId of the site where the asset publisher instance is placed. This change is safe because in this context the scopeGroupId and the siteGroupId are always going to be the same (which is the same reason why it has worked so far). But semantically speaking what makes sense is the siteGroupid.