LPS-33005 Replace all references to journalTemplatesPath by generic templatesPath
The journalTemplatesPath had the format _JOURNAL_SEPARATOR_/companyId/groupId/templateId. The journalTemplate.templateId was not unique,  so companyId and groupId were required. However, since the groupId is that of the page were the template is being rendered, the journalTemplatesPath couldn't be used for including templates from other groups (e.g. global templates). The new templatesPath format is _TEMPLATES_SEPARATOR/templateId. The ddmTemplate.templateId is unique, so companyId and groupId is no longer required. Furthermore, the templatesPath can now be used for including global templates.