LPS-54156 Remove dependency to StringBundler and replace it by StringBuilder.
This is necessary because this module is supposed to NOT depend on any portal module and if we use StringBundler we need to copy a lot of classes internally. Otherwise, as it is right now it is not working because we are already copying ReflectionUtil from portal-service and that makes BND ignoring the com.liferay.portal.kernel.util package in imports section of MANIFEST, thus this module doesn't see any StringBundler.